+++ eleventy.config.js
@@ -1,17 +1,23 @@
 import syntaxHighlight from "@11ty/eleventy-plugin-syntaxhighlight";
 
-export default function (eleventyConfig) {
-  eleventyConfig.addPlugin(syntaxHighlight);
-  eleventyConfig.addPassthroughCopy("docs-src/docs.css");
-  eleventyConfig.addPassthroughCopy("public/");
-  eleventyConfig.addPassthroughCopy("docs-src/public/");
-  eleventyConfig.addPassthroughCopy("dist/");
-  eleventyConfig.addPassthroughCopy("assets/");
+export default function (config) {
+  config.addPlugin(syntaxHighlight);
+  config.addPassthroughCopy("docs-src/docs.css");
+  config.addPassthroughCopy("public/");
+  config.addPassthroughCopy("docs-src/public/");
+  config.addPassthroughCopy("dist/");
+  config.addPassthroughCopy("assets/");
 
-  eleventyConfig.addPassthroughCopy("node_modules/prismjs/themes/prism-okaidia.min.css");
-  eleventyConfig.addPassthroughCopy("node_modules/prismjs/components/prism-core.min.js");
-  eleventyConfig.addPassthroughCopy("node_modules/prismjs/components/prism-markup.min.js");
-  eleventyConfig.addPassthroughCopy("node_modules/prismjs/plugins/autoloader/prism-autoloader.min.js");
+  // we don't use the node_modules folder in the docs deployment because Netlify
+  // will refuse to serve the node_modules/ directory through their web server
+  // to get around this, we copy the necessary files from node_modules to a
+  // deps/ directory
+  config.addPassthroughCopy({ "node_modules/prismjs/themes/prism-okaidia.min.css": "deps/prism-okaidia.min.css" });
+  config.addPassthroughCopy({ "node_modules/prismjs/components/prism-core.min.js": "deps/prism-core.min.js" });
+  config.addPassthroughCopy({ "node_modules/prismjs/components/prism-markup.min.js": "deps/prism-markup.min.js" });
+  config.addPassthroughCopy({
+    "node_modules/prismjs/plugins/autoloader/prism-autoloader.min.js": "deps/prism-autoloader.min.js",
+  });
 
   return {
     dir: {